pyproject.toml
[tool.poetry]
name = "lagransala"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.8"}
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlmodel = "^0.0.22"
aiohttp = {extras = ["speedups"], version = "^3.11.12"}
beautifulsoup4 = "^4.13.3"
markdownify = "^0.14.1"
pyyaml = "^6.0.2"
instructor = "^1.7.2"
anthropic = "^0.45.2"
jinja2 = "^3.1.5"
babel = "^2.17.0"
langfuse = "^2.59.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
pytest = "^8.3.4"
black = "^25.1.0"
isort = "^6.0.0"
flake8 = "^7.1.1"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/black_compatibility/
profile = "black"
